+++ PS7-x64/Modules/AdminToolbox/ChangeLog.md
@@ -130,3 +130,4 @@
 * **7.1.4.0**  Updated multiple module dependencies
 * **8.0.0.0**  Moved two functions to the file management module. Added cmder to the module.
 * **8.0.1.0**  Updated multiple module dependencies
+* **8.1.0.0**  Updated multiple module dependencies